Changes based on PR feedback
